requirements.txt
eyed3==0.9.7
fastapi==0.111.0
python-dotenv==1.0.1
pytube==15.0.0
